backends to release branch

+++ sdonboarding.Server/Controller/SalesController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using sdonboarding.Server.Models;
 
 namespace sdonboarding.Server.Controller
@@ -11,10 +12,12 @@ namespace sdonboarding.Server.Controller
     {
 
         private readonly OnBoardingContext _context;
+        private readonly ILogger<SalesController> _logger;
 
-        public SalesController(OnBoardingContext context)
+        public SalesController(OnBoardingContext context, ILogger<SalesController> slogger)
         {
             _context = context;
+            _logger = slogger;
         }
 
 
@@ -22,15 +25,23 @@ namespace sdonboarding.Server.Controller
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Dtos.SalesDto>>> GetSales()
         {
-            var _sales = await _context.Sales.Select(s => Mappers.SalesMapper.EntityToDto(s)).ToListAsync();
+            try
+            {
+                var _sales = await _context.Sales.Select(s => Mappers.SalesMapper.EntityToDto(s)).ToListAsync();
 
-            if (_sales.Count > 0)
-            {
-                return Ok(_sales);
+                if (_sales.Count > 0)
+                {
+                    return Ok(_sales);
+                }
+                else
+                {
+                    return BadRequest("There are no Sales at the moment");
+                }
             }
-            else
+            catch (Exception e)
             {
-                return BadRequest("There are no Sales at the moment");
+                _logger.LogError(e, "An error occurred while processing your request.");
+                return StatusCode(500, "An error occurred while processing your request.");
             }
         }
 
@@ -39,14 +50,30 @@ namespace sdonboarding.Server.Controller
         [HttpGet("{id}")]
         public async Task<ActionResult<Dtos.SalesDto>> GetSales(int id)
         {
-            var sales = await _context.Sales.FindAsync(id);
-
-            if (sales == null)
+            if (id <= 0)  // Check if id is invalid (0 or negative)
             {
-                return NotFound();
+                return BadRequest("Invalid store ID.");
             }
+            else
+            {
 
-            return Mappers.SalesMapper.EntityToDto(sales);
+                try
+                {
+                    var sales = await _context.Sales.FindAsync(id);
+
+                    if (sales == null)
+                    {
+                        return NotFound();
+                    }
+
+                    return Mappers.SalesMapper.EntityToDto(sales);
+                }
+                catch (Exception e)
+                {
+                    _logger.LogError(e, "An error occurred while processing your request.");
+                    return StatusCode(500, "An error occurred while processing your request.");
+                }
+            }
         }
 
 
@@ -54,29 +81,56 @@ namespace sdonboarding.Server.Controller
         [HttpPost]
         public async Task<ActionResult<Sale>> PostSales(Dtos.SalesDto sales)
         {
-            var entity = Mappers.SalesMapper.DtoToEntity(sales);
+            try
+            {
+                var entity = Mappers.SalesMapper.DtoToEntity(sales);
 
-            _context.Sales.Add(entity);
+                _context.Sales.Add(entity);
 
-            await _context.SaveChangesAsync();
+                await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetSales", new { id = sales.Id }, Mappers.SalesMapper.EntityToDto(entity));
+                return CreatedAtAction("GetSales", new { id = sales.Id }, Mappers.SalesMapper.EntityToDto(entity));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "An error occurred while processing your request.");
+                return StatusCode(500, "An error occurred while processing your request.");
+            }
         }
 
         // DELETE: api/Sales/7
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteSales(int id)
         {
-            var sale = await _context.Sales.FindAsync(id);
-            if (sale == null)
+
+            if (id <= 0)  // Check if id is invalid (0 or negative)
             {
-                return NotFound();
+                return BadRequest("Invalid store ID.");
             }
+            else
+            {
+                try
+                {
 
-            _context.Sales.Remove(sale);
-            await _context.SaveChangesAsync();
+                    var sale = await _context.Sales.FindAsync(id);
+                    if (sale == null)
+                    {
+                        return NotFound();
+                    }
 
-            return NoContent();
+                    _context.Sales.Remove(sale);
+                    await _context.SaveChangesAsync();
+
+                    return NoContent();
+                }
+                catch (Exception e)
+                {
+                    _logger.LogError(e, "An error occurred while processing your request.");
+                    return StatusCode(500, "An error occurred while processing your request.");
+
+                }
+
+            }
         }
